Renombrado el usuario por worker  * Debido a un mal diseño de la base de datos, se tuvo que modificar esto
