+++ controllers/index.js
@@ -1,5 +1,5 @@
 exports.clientCtrl = require('./clientCtrl');
-exports.userCtrl = require('./userCtrl');
+exports.workerCtrl = require('./workerCtrl');
 exports.listEmailCtrl = require('./listEmailCtrl');
 exports.wishListCtrl = require('./wishListCtrl');
 exports.cartCtrl = require('./cartCtrl');
